Add files via upload
With this update, I adjusted the settlement() parameters to relax the self-extinction parameters. The results are a mix of replicates with some showing survival and growth in numbers beyond 1000 years, some show diminishing numbers but with survival beyond 1000 years as well and finally, some show extinction but only after many years (e.g 500+ years)

You can change the seed value on the code "RSsim(seed = 13976)" or increase the number of replicates at:

"Simulation(Simulation = 1,
                  Years = 1000,
                  Replicates = 10,
                  OutIntRange = 5,
                  OutIntPop = 5,
                  OutIntOcc = 5)"

to see what different simulations result in.