Add files via upload
Hello, I have been working on the simulation with the changes to reflect the introduction points in australia correctly and therefore simulate range dispersal foe Australia only. i have not been able to share relevant output yet and  i appologise for the delays. This is the challenge I am facing:

There is a particluar file format required for the InitIndsFile for it to work. From my observations, the data is in the right format and we even get "Initial individuals file OK:/..." when we run the simulation.

There is however an error that "*** Error in InitIndsFile at line 1: Species must be 0"

this is despite the values for Species being 0 in the file. I have shared the code file (reslts maxent2.rmd) and the data (inputs folder) here. I cant seem to figure out why the error occurs despite this.

I have also not gotten a response addressing this from the rangeshiftr github discussion page. I doont know if there is somrething you may be able to see in the data file that may be causing the problem? This is the issue delaying the outputs from the simulation.